# backend/services/gateway/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: cportal-lrfe API (through the gateway)
  version: 0.1.0
  description: >-
    Every endpoint the browser can reach, at its /api/... path. The gateway only forwards requests: each service checks the token and
    permissions itself. Service-only endpoints are listed too, but answer 401 to a user token.


    In development, `ng serve` proxies /api to the gateway (angular-app/proxy.conf.json), so the paths work from the app's origin as they
    are.
servers:
  - url: http://localhost:3500
    description: gateway (local development)
...
paths:
...
  /api/records:
    get:
      tags:
        - land-records · Records
      summary: List and search land records
      description: >-
        `{ items, total }`. `q` searches the record number, the parcel, owners' names and ID numbers, and document references. `status`:
        `draft` (never committed), `committed`, `in_review` (a version waits for approval), `needs_review` (a linked document was
        corrected).


        **Access:** Signed-in user with permission `record.view`.
      operationId: land-records_getRecords
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - draft
              - committed
              - in_review
              - needs_review
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.view
    post:
      tags:
        - land-records · Records
      summary: Create a land record (draft v1)
      description: >-
        `{ parcel }`, or `{ edrmsDocumentId }` to read the parcel from a filed document's property field and link that document. One record
        per parcel: 409 with `recordId` when the parcel already has one.


        **Access:** Signed-in user with permission `record.create`.
      operationId: land-records_postRecords
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                parcel:
                  type: object
                attributes:
                  type: object
                edrmsDocumentId:
                  type: string
                  format: uuid
              anyOf:
                - required:
                    - parcel
                - required:
                    - edrmsDocumentId
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.create
  /api/records/{id}:
    get:
      tags:
        - land-records · Records
      summary: One land record with its current version and open draft
      description: "**Access:** Signed-in user with permission `record.view`."
      operationId: land-records_getRecordsId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.view
  /api/records/{id}/comments:
    get:
      tags:
        - land-records · Records
      summary: Comments on a record
      description: "**Access:** Signed-in user with permission `record.view`."
      operationId: land-records_getRecordsIdComments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.view
    post:
      tags:
        - land-records · Records
      summary: Comment on a record
      description: |-
        `{ body }`. Kept with the version open at the time.

        **Access:** Signed-in user with permission `record.comment`.
      operationId: land-records_postRecordsIdComments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - body
              properties:
                body:
                  type: string
                  minLength: 1
                  maxLength: 4000
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.comment
  /api/records/{id}/draft:
    patch:
      tags:
        - land-records · Records
      summary: "Edit the draft: core fields, attributes, suggestions"
      description: >-
        `{ revision, changes, accept, reason }`. `changes`: parcel (same kind), extent, tenure, owners, encumbrances, attributes (null
        removes). `accept`: take over the documents' suggested owners, extent or encumbrances. Owners or an extent that differ from the
        suggestions are marked as entered by hand and need a `reason`. Suggestions (`derived`) and checks are recomputed. 409 when
        `revision` is not the draft's current revision.


        **Access:** Signed-in user with any of `record.create`, `record.link`.
      operationId: land-records_patchRecordsIdDraft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - revision
              properties:
                revision:
                  type: integer
                  minimum: 1
                changes:
                  type: object
                accept:
                  type: array
                  uniqueItems: true
                  items:
                    type: string
                    enum:
                      - owners
                      - extent
                      - encumbrances
                reason:
                  type: string
                  minLength: 1
                  maxLength: 1000
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        anyOf:
          - record.create
          - record.link
    post:
      tags:
        - land-records · Records
      summary: Open a new draft to change a committed record
      description: >-
        The draft vN+1 is a copy of the current version, which stays current until the new one is approved. One open draft per record (409).
        The record then carries `draftDiff`: what the draft changes against the current version.


        **Access:** Signed-in user with any of `record.create`, `record.link`.
      operationId: land-records_postRecordsIdDraft
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        anyOf:
          - record.create
          - record.link
  /api/records/{id}/draft/documents:
    post:
      tags:
        - land-records · Records
      summary: Link a filed document to the draft
      description: >-
        `{ revision, edrmsDocumentId }`. The document is pinned at its current EDRMS version, with its seal and a snapshot of its verified
        fields.


        **Access:** Signed-in user with permission `record.link`.
      operationId: land-records_postRecordsIdDraftDocuments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - revision
                - edrmsDocumentId
              properties:
                revision:
                  type: integer
                  minimum: 1
                edrmsDocumentId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.link
  /api/records/{id}/draft/documents/{edrmsDocumentId}:
    delete:
      tags:
        - land-records · Records
      summary: Remove a document from the draft
      description: |-
        `?revision=` the draft's current revision.

        **Access:** Signed-in user with permission `record.unlink`.
      operationId: land-records_deleteRecordsIdDraftDocumentsEdrmsDocumentId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: edrmsDocumentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: revision
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.unlink
  /api/records/{id}/draft/documents/{edrmsDocumentId}/refresh:
    post:
      tags:
        - land-records · Records
      summary: Adopt a document's newer EDRMS version
      description: >-
        `{ revision }`. After a correction in edrms: the pinned version, seal and fields are updated in the draft, suggestions and checks
        recomputed. Goes through review like any change.


        **Access:** Signed-in user with permission `record.link`.
      operationId: land-records_postRecordsIdDraftDocumentsEdrmsDocumentIdRefresh
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: edrmsDocumentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - revision
              properties:
                revision:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.link
  /api/records/{id}/draft/submit:
    post:
      tags:
        - land-records · Records
      summary: Submit the draft for review
      description: >-
        `{ revision }`. Refused (400) while the data is incomplete (`problems`) or an error check fails (`failing`). The draft is frozen and
        the bpm process `land-record-review` starts: an approval task for `record.finalize`, never the submitter.


        **Access:** Signed-in user with permission `record.link`.
      operationId: land-records_postRecordsIdDraftSubmit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
                - revision
              properties:
                revision:
                  type: integer
                  minimum: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.link
  /api/records/{id}/draft/withdraw:
    post:
      tags:
        - land-records · Records
      summary: Withdraw a version from review
      description: |-
        Only the submitter, before a decision. The review task disappears and the draft is editable again.

        **Access:** Signed-in user with permission `record.link`.
      operationId: land-records_postRecordsIdDraftWithdraw
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.link
  /api/records/{id}/history:
    get:
      tags:
        - land-records · Records
      summary: "History: every committed version"
      description: >-
        `{ intact, flags, versions }`, newest first: submitter, approver, dates, review comment, seal with `intact` and `linked`, and `diff`
        against the version before.


        **Access:** Signed-in user with permission `record.view`.
      operationId: land-records_getRecordsIdHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.view
  /api/records/{id}/review:
    get:
      tags:
        - land-records · Records
      summary: What the reviewer sees
      description: >-
        `{ record, version, checks, blocking, overrides, diff, currentVersion }`. `diff` against the current committed version: `fields`
        (path, before, after), `owners` (added, removed, changed), `encumbrances`, `documents` (added, removed, updated to a newer EDRMS
        version).


        **Access:** Signed-in user with permission `record.view`.
      operationId: land-records_getRecordsIdReview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "400":
          description: Invalid request (failed validation)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing, invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Not permitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
        - bearerAuth: []
      x-access:
        allOf:
          - record.view
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from POST /api/auth/login or /api/auth/mfa (15 minutes; renew with POST /api/auth/refresh).
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Service-to-service token (typ "service"), signed with the shared JWT_SECRET. Not available to browsers.
    refreshCookie:
      type: apiKey
      in: cookie
      name: lrfe_rt
      description: httpOnly refresh cookie, path /api/auth, set at sign-in.
  schemas:
    Error:
      type: object
      required:
        - status
        - message
        - timestamp
        - path
      properties:
        status:
          type: string
          example: Forbidden
          description: HTTP status text
        message:
          type: string
          example: Not permitted
        details:
          description: Extra data on some errors, e.g. `{ perm }` on 403, `{ blockers }` when a document cannot be filed yet
        timestamp:
          type: string
          format: date-time
        path:
          type: string
          example: /users